One task left

+++ app/models.py
@@ -12,7 +12,7 @@ class Subscription(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     user_id = db.Column(db.Integer)
     name = db.Column(db.String(80))
-    amount = db.Column(db.Integer)
+    amount = db.Column(db.Float)
     billing_cycle = db.Column(db.String(10))
     next_due_date = db.Column(db.DateTime)
